refactor: JWT 모듈 패키지 구조 개선
--- 본문 ---
- JwtToken 클래스를 dto 패키지로 이동하여 DTO 역할을 명확히 했습니다.
- JwtRegistry 및 CaffeineJwtRegistry 클래스를 registry 패키지로 이동하여 레지스트리 관련 로직을 그룹화했습니다.
- JwtProperties 클래스를 config 패키지로 이동하여 설정 관련 클래스를 통합했습니다.
- 관련 클래스들의 패키지 위치를 재조정하여 모듈의 응집도를 높이고 가독성을 개선했습니다.
