+++ src/main/java/io/github/metdaisy/amaazon/global/security/jwt/config/JwtConfig.java
@@ -1,6 +1,5 @@
 package io.github.metdaisy.amaazon.global.security.jwt.config;
 
-import io.github.metdaisy.amaazon.global.security.jwt.JwtProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
