+++ src/main/java/io/github/metdaisy/amaazon/global/security/jwt/config/JwtProperties.java
@@ -1,4 +1,4 @@
-package io.github.metdaisy.amaazon.global.security.jwt;
+package io.github.metdaisy.amaazon.global.security.jwt.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
